# swagger.yaml
openapi: 3.0.3
info:
  title: CargoFlow Logistics API
  version: 1.0.0
  description: >
    API documentation for Shipment Management System.

    USER STORIES & API MAPPING:

    Auth:
    - Register → User creates account
    - Login → User accesses system
    - Profile → User views/updates details

    Admin:
    - Dashboard → Monitor system
    - Shipments → Manage shipments
    - Agents → Manage delivery agents
    - Clients → Manage business clients

    Client:
    - Shipments → Create & view shipments
    - Tracking → Track deliveries
    - Business → Manage business profile

    Agent:
    - Dashboard → View assigned tasks
    - Location → Update delivery location

    Chatbot:
    - Chat → Ask shipment-related queries
    - Health → Check chatbot status
    - Rebuild → Refresh knowledge base
servers:
  - url: http://localhost:8000
...
paths:
...
  /api/v1/admin/dashboard:
    get:
      tags: [Admin]
      summary: Admin Dashboard
      description: "User Story: As an admin, I want to view overall system metrics."
      security:
        - bearerAuth: []
      responses:
        "200": { description: Dashboard data }
  /api/v1/admin/dashboard/shipments:
    get:
      tags: [Admin]
      summary: Shipment Stats
      description: "User Story: As an admin, I want to view shipment statistics."
      security:
        - bearerAuth: []
      responses:
        "200": { description: Shipment data }
  /api/v1/admin/dashboard/agents:
    get:
      tags: [Admin]
      summary: Agent Stats
      description: "User Story: As an admin, I want to view agent performance."
      security:
        - bearerAuth: []
      responses:
        "200": { description: Agent data }
  /api/v1/admin/dashboard/clients:
    get:
      tags: [Admin]
      summary: Client Stats
      description: "User Story: As an admin, I want to view client data."
      security:
        - bearerAuth: []
      responses:
        "200": { description: Client data }
  /api/v1/admin/shipments:
    post:
      tags: [Admin]
      summary: Create Shipment
      description: "User Story: As an admin, I want to create shipments."
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminShipmentCreate'
      responses:
        "201": { description: Created }
  /api/v1/admin/shipments/{shipment_id}/assign/{agent_id}:
    post:
      tags: [Admin]
      summary: Assign Agent
      description: "User Story: As an admin, I want to assign agents."
      security:
        - bearerAuth: []
      parameters:
        - name: shipment_id
          in: path
          required: true
          schema: { type: integer }
        - name: agent_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Assigned }
...
  /api/v1/admin/business_clients:
    post:
      tags: [Admin]
      summary: Add Client
      description: "User Story: As an admin, I want to add business clients."
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegister'
      responses:
        "201": { description: Client added }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    UserRegister:
      type: object
      properties:
        name: { type: string }
        email: { type: string }
        password: { type: string }
...
    AdminShipmentCreate:
      type: object
...